Add dedicated class for parameter storage. Again implement it at the example of the base_station. Move decorator to new class
